+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add 'enable_cookies_banner' option to module config file
+
 <!-- markdownlint-disable line-length -->
 
 ## [v-egi-2.0.4](https://github.com/EGI-Foundation/simplesamlphp-module-themeegi/compare/v-egi-2.0.2...v-egi-2.0.4) - 2021-07-12